Built out the basics of the entity
Entity is a new concept that sits a level higher than what we've had in
the past with Content. Entity defines the overall structure of the
object -- meaning both the content and the associated objects.

This commit adds the basic structure of the Entity as it pertains to
Content, as well as the mechanism for creating it.

It will likely lead to the destruction of a few other concepts like:

- IlluminatedObjects
- The Illuminated service
- Object heads
- The object manager service